requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.10.0
Flask-Limiter==3.7.0
Flask-Talisman==1.1.0
Flask-Bcrypt==1.0.1
APScheduler==3.10.4
itsdangerous==2.2.0
requests==2.32.3
python-dotenv==1.0.1
email-validator==2.2.0
WTForms==3.1.2
